scan_server/requirements.txt
fastapi
pydantic_settings
python-gvm==23.10.0
zaproxy==0.2.0
gunicorn